polarcam_live/requirements.txt
numpy
opencv-python
Pillow
scipy
matplotlib
